modprobe: Fix format string bug in do_command()
gcc produces the following warning on x86_64:

modprobe.c:786: warning: field precision should have type 'int', but argument 3 has type 'long int'

This bug is currently harmless on x86_64, because parameters are
passed in registers, but could cause misbehavior on other 64-bit
platforms.

+++ modprobe.c
@@ -782,7 +782,7 @@ static void do_command(const char *modname,
 	while ((p = strstr(replaced_cmd, "$CMDLINE_OPTS")) != NULL) {
 		char *new;
 		asprintf(&new, "%.*s%s%s",
-			 p - replaced_cmd, replaced_cmd, cmdline_opts,
+			 (int)(p - replaced_cmd), replaced_cmd, cmdline_opts,
 			 p + strlen("$CMDLINE_OPTS"));
 		NOFAIL(new);
 		free(replaced_cmd);
